fstests: split kiling pids vs systems on hang
The decision to reset systems is different than killing
associated tasks and sending a report of a hang. In practice
we actually want to *currently* manually review a hang.
And so add a new kconfig option to specicy the kill. We
leave the old kconfig entry in place for the reset.

While at it, fix having teh run_kernel_ci script to only
email out the diff file, as that is the short version of the
file. Right now we are emailing the full report, and we don't
care to see the ansible task output on the email report.

If the run time files are empty, don't print that
information out.

Lastly, add support to list pids, so that way you can
safely scrape to see what PIDs would be killed if you
issue the command, or you can check if the kill script
actually did what it was supposed to.

+++ scripts/workflows/fstests/kill_pids.sh
@@ -5,9 +5,27 @@ fi
 source ${TOPDIR}/.config
 source ${TOPDIR}/scripts/lib.sh
 
+CALL=$(basename $0)
+LIST_ONLY="False"
+
+if [[ "$CALL" == "list_pids.sh" ]]; then
+	LIST_ONLY="true"
+fi
+
 FS=$CONFIG_FSTESTS_FSTYP
 STRING='CONFIG_FSTESTS_FSTYP='
 PID_LIST=$(ps -fu $USER | awk '{print $2}')
+
+list_pid()
+{
+	if [[ "$2" == "kill" ]]; then
+		echo "Would be killed:"
+	elif [[ "$2" == "skip" ]]; then
+		echo "Would be skipped:"
+	fi
+	cat /proc/$1/cmdline
+	echo
+}
 
 for i in $PID_LIST; do
 	if [[ ! -d /proc/$i ]]; then
@@ -24,30 +42,51 @@ for i in $PID_LIST; do
 
 	grep -q run_kernel_ci /proc/$i/cmdline
 	if [[ $? -eq 0 ]]; then
+		if [[ "$LIST_ONLY" == "true" ]]; then
+			list_pid $i skip
+			continue
+		fi
 		continue
 	fi
 
 	grep -q "fstests-baseline-loop" /proc/$i/cmdline
 	if [[ $? -eq 0 ]]; then
+		if [[ "$LIST_ONLY" == "true" ]]; then
+			list_pid $i skip
+			continue
+		fi
 		continue
 	fi
 
 	grep -q make /proc/$i/cmdline
 	if [[ $? -eq 0 ]]; then
+		if [[ "$LIST_ONLY" == "true" ]]; then
+			list_pid $i skip
+			echo "Would be killed:"
+			cat /proc/$i/cmdline
+			echo
+			continue
+		fi
 		kill -SIGTERM -- -${i} 2>/dev/null
 		kill -SIGTERM ${i} 2>/dev/null
 	fi
 
 	grep -q run_loop /proc/$i/cmdline
 	if [[ $? -eq 0 ]]; then
+		if [[ "$LIST_ONLY" == "true" ]]; then
+			list_pid $i kill
+			continue
+		fi
 		kill -SIGTERM -- -${i} 2>/dev/null
 		kill -SIGTERM ${i} 2>/dev/null
 	fi
 
 	grep -q ansible-playbook /proc/$i/cmdline
 	if [[ $? -eq 0 ]]; then
-		cat /proc/$i/cmdline
-		echo
+		if [[ "$LIST_ONLY" == "true" ]]; then
+			list_pid $i kill
+			continue
+		fi
 		kill -SIGTERM -- -${i} 2>/dev/null
 		kill -SIGALRM -- -${i} 2>/dev/null
 		kill -SIGALRM ${i} 2>/dev/null
@@ -56,6 +95,10 @@ for i in $PID_LIST; do
 
 	grep -q ssh /proc/$i/cmdline
 	if [[ $? -eq 0 ]]; then
+		if [[ "$LIST_ONLY" == "true" ]]; then
+			list_pid $i kill
+			continue
+		fi
 		kill -SIGTERM -- -${i} 2>/dev/null
 		kill -SIGTERM ${i} 2>/dev/null
 	fi
